Work in progress

+++ src/framerate.cpp
@@ -11,7 +11,7 @@ void framerate::loop(float_t delta) {
   _start = now;
 
   if (_elapsed >= 1000) {
-    fmt::println("{:.1f}", static_cast<float>(_frames) * 0.001f * static_cast<float>(_elapsed));
+    fmt::println("{:.1f}", _frames * 0.001f * _elapsed);
 
     _elapsed = 0;
     _frames = 0;
